adedpage for done widgets
