+++ lib/Redesign/buildcomponents.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'RecipePages/ComponentsRedesign.dart';
 import 'RecipePages/imagesclass.dart';
+import 'doneshowcase.dart';
 
 class BuildComponentPage extends StatelessWidget {
   const BuildComponentPage({Key? key}) : super(key: key);
@@ -54,93 +55,6 @@ class BuildComponentPage extends StatelessWidget {
           ),
         ),
       ),
-    );
-  }
-}
-
-class RowOfIcons80 extends StatelessWidget {
-  const RowOfIcons80({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return const Column(
-      children: [
-        Row(
-          children: [
-            IconTile(iconName: 'Pig100'),
-            IconTile(iconName: 'Soup100'),
-            IconTile(iconName: 'SeaFood100'),
-            IconTile(iconName: 'Fruit100'),
-            IconTile(iconName: 'Drinks100'),
-            IconTile(iconName: 'Bread100'),
-            IconTile(iconName: 'Lamb100'),
-            IconTile(iconName: 'Cow100'),
-            IconTile(iconName: 'Turkey100'),
-            IconTile(iconName: 'Dessert100'),
-            IconTile(iconName: 'Chicken100'),
-            IconTile(iconName: 'Fish100'),
-            IconTile(iconName: 'Vegetable100'),
-          ],
-        ),
-        Row(
-          children: [
-            IconTileText(
-              iconName: 'Pig100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Soup100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'SeaFood100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Fruit100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Drinks100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Bread100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Lamb100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Cow100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Turkey100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Dessert100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Chicken100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Fish100',
-              iconTextPairs: iconTextPairs,
-            ),
-            IconTileText(
-              iconName: 'Vegetable100',
-              iconTextPairs: iconTextPairs,
-            ),
-          ],
-        ),
-      ],
     );
   }
 }
